+++ model_fetch_firestore/lib/src/macros/firestore_model/filter_param_visitors/abstract.dart
@@ -1,0 +1,9 @@
+import '../filter_param.dart';
+
+abstract class FilterParamVisitor<R> {
+  const FilterParamVisitor();
+
+  R visit(FilterParam param) => param.accept(this);
+
+  R visitEquals(EqualsFilterParam param);
+}